+++ RELEASENOTES.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## Unreleased
+
+- PSR4 namespacing. Now in `Addwiki\Mediawiki\Api`
+- Client classes, such as `MediaWikiApi`, moved to `Addwiki\Mediawiki\Api`
+
 ## Version 2.8 (16 February 2021)
 
 - Installable with 7.3+ (including PHP8)